Tensorflow Dataset for GTA
This patch defines a dataset as a tensorflow dataset, which
downloads an archive from Mendeley and converts it to Tensor.
