+++ Pipfile
@@ -15,6 +15,7 @@ keract = ">=4.4.0"
 h5py = "==2.10.0"
 Keras = ">=2.4.3"
 tensorflow = ">=2.4.1"
+tensorflow-datasets = ">=4.4.0"
 pyts = ">=0.11.0"
 keras-rl2 = "==1.0.4"
 pytest = "*"
