+++ deeplookup/datasets/__init__.py
@@ -1,0 +1,1 @@
+from deeplookup.datasets.core import *  # noqa
