--- deeplookup/weights/b-dga-ybubnov.h5
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:25e7ba9182969d32d5bc3ddb62957f9a2919e752e584f0932d103e4721d1cda3
-size 6358776
